backend_v2/requirements.txt
# Django Core
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter
django-environ==0.11.2    # Environment variables

# Image processing
Pillow==11.3.0

# Admin interface
django-admin-interface==0.30.0
django-colorfield==0.9.0

# Background tasks (optional)
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# File storage
boto3==1.35.80
django-storages==1.14.4

# SMS service
requests==2.32.3

# Push Notifications (Firebase Admin SDK)
firebase-admin==6.5.0

# Payment processing
stripe==10.8.0

# Monitoring and logging
sentry-sdk==2.18.0

# Development tools
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Security
django-ratelimit==4.1.0
django-ratelimit2==0.4.2

# Utilities
python-decouple==3.8
django-filter==24.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
